pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-dj"
version = "0.1.0"
description = "An agentic DJ system — dissertation thesis project"
requires-python = ">=3.11"
dependencies = [
    "spotipy",
    "pylast",
    "python-dotenv",
    "anthropic",
    "openai",
    "google-generativeai",
    "langchain",
    "langgraph",
    "music21",
    "streamlit",
    "pydantic",
    "requests",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_dj"]
